Add files via upload
We are able to complete our iteration1 by the requirements. Where user has to run the program the user is also is also responsible in collecting all the datas.The functional requirements will show what should the software system do but this are specified by the user while the non functional requirements how should the software system fulfil the functional requirements so it concentrate on the users expectation.